Delete Auth added

+++ routes/userRoutes/posts/deletePost.js
@@ -9,18 +9,31 @@ module.exports = router.post(
   async (req, res) => {
     try {
       const { id } = req.params;
+      const user = req.user;
 
-      // create Post
-      await Post.deleteOne({ _id: id }).then((data) => {
-        if (data) {
-          res.status(200).json({
-            data: {
-              status: true,
-              message: "Post deleted successfully!",
-            },
-          });
-        }
-      });
+      // if the post is in the collection
+      const postExist = await Post.findById(id).then((data) => data);
+
+      if (postExist?.userId.toString() === user?._id.toString()) {
+        // Delete Post
+        return await Post.deleteOne({ _id: id }).then((data) => {
+          if (data) {
+            res.status(200).json({
+              data: {
+                status: true,
+                message: "Post deleted successfully!",
+              },
+            });
+          }
+        });
+      } else {
+        return res.status(401).json({
+          data: {
+            status: true,
+            message: "Post deleted unsuccessfully!",
+          },
+        });
+      }
     } catch (error) {
       console.log(error);
       res.status(400).json({
